refactor(daengle): refactor folder structure
