--- apps/daengle/src/pages/onboarding/hooks/index.ts
@@ -1,2 +1,0 @@
-export { default as useValidatePetForm } from './use-validate-pet-form';
-export { default as useValidateUserForm } from './use-validate-user-form';
