--- apps/daengle/src/pages/onboarding/pages/index.ts
@@ -1,3 +1,0 @@
-export { default as PetInfo } from './pet-info';
-export { default as SearchAddress } from './search-address';
-export { default as UserInfo } from './user-info';
